Disable SSH connection sharing because it breaks the build process
The `-S` option with the parameter `none` disable connection sharing
for SSH. The problem is pipeing fails with the spawned ssh connection
because of some timeout I think.

If you disable connection sharing for the SSH connection everything
is working fine. This is only required if you use ControlMaster in
your ~/.ssh/config file.