Set default network to one public network and allow the user the option to overwrite requirements
This should fix #7. It allow the user to overwrite all requirements based on the `manifest.json`. If no `requirements.networks` are set we use the default of one network interface. If `minram` is used it will be added to the requirements as well but could be overwritten by the `requirements` list.